backend/src/requirements.txt
asgiref==3.2.3
autopep8==1.5
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
